requirement.txt
pandas
matplotlib
numpy
openpyxl
scikit-learn
torchvision
torchsummary
tensorboard